// components/home/blogs/types/data.ts
export type BlogHeroData = {
  title: string;
  description: string;
  image: string;
  imageAlt: string;
  badgeText: string;
};

export type FeaturedBlogData = {
  eyebrow: string;
  category: string;
  title: string;
  description: string;
  tags: string[];
  image: string;
  imageAlt: string;
};